src/database.py
import mysql.connector
import os

def get_db_connection():
    host = os.environ.get("MYSQL_HOST"),
    user = os.getenv("MYSQL_USER")
    password = os.getenv("MYSQL_PASSWORD")
    database = os.getenv("MYSQL_DATABASE")

    connection = mysql.connector.connect(
        host=host,
        user=user,
        password=password,
        database=database
    )
    return connection