src/requirements.txt
uvicorn==0.32.0
fastapi==0.115.3
starlette==0.41.0
anyio==4.6.2.post1
kafka-python @ git+https://github.com/dpkp/kafka-python.git@a6d0579d3cadd3826dd364b01bc12a2173139abc
colorlog==6.8.2
aiokafka==0.12.0
mysql-connector-python==9.1.0
pytz==2024.2